+++ app/components/BasicMap.jsx
@@ -17,7 +17,7 @@ require("openlayers/css/ol.css");
 import styles from '../styles/map.css';
 import {Icon} from 'react-mdc-web';
 import {getAttribution} from '../utils/mapUtils';
-import type {Sensors, MapProps, BasicMapState} from '../utils/flowtype';
+import type {MapProps, BasicMapState} from '../utils/flowtype';
 
 class BasicMap extends Component {
     state: BasicMapState;
@@ -51,6 +51,7 @@ class BasicMap extends Component {
 
         return (
          <div>
+
             <div id='map' className={styles.root}>
             </div>
 
@@ -87,6 +88,7 @@ class BasicMap extends Component {
                 <button id="drawClearButton" title="Click to Reset Drawing Selection">
                     <Icon name="clear"/></button>
             </div>
+
         </div>);
 
     }
@@ -226,6 +228,9 @@ class BasicMap extends Component {
                 // This is the button that will reset the points
                 that.props.selectShapeLocation(selectPointsLocations, drawCoordinates);
 
+                // Reset the State Variable
+                that.setState({customLocationFilterVectorExtent: []});
+
                 let keep_draw_active = document.getElementById('draw');
                 if (keep_draw_active) {
                     keep_draw_active.click();
@@ -348,7 +353,6 @@ class BasicMap extends Component {
                 let drawCircle = new ol.interaction.Draw({
                     type: 'Circle',
                     source: customLocationFilterVector,
-                    geometryFunction: ol.interaction.Draw.createRegularPolygon(40),
                 });
 
                 theMap.addInteraction(drawCircle);
@@ -399,8 +403,15 @@ class BasicMap extends Component {
                         }
                     }
                     // Get the shape coordinates
-                    let drawCoordinates = drawExtent.getCoordinates();
-
+                    // (1) Get the Units for the Map Projection
+                    let units = theMap.getView().getProjection().getUnits();
+                    // (2) Get the Center Point of the Circle
+                    let drawCenter = drawExtent.getCenter();
+                    // (3) Get the Radius of the Circle in Meters
+                    let drawRadius = (drawExtent.getRadius() * ol.proj.METERS_PER_UNIT[units])/1000;
+                    // Assemble the Coordinates for the API call
+                    let drawCoordinates = [drawCenter.concat(drawRadius)];
+                    // Call the appropriate function
                     that.props.selectShapeLocation(selectPointsLocations, drawCoordinates);
 
                 });