+++ app/reducers/chosenTrends.js
@@ -4,11 +4,11 @@
 
 
 import {
-    ADD_ANALYSIS_COUNT, ADD_CHOOSE_TRENDS, ADD_REGION_DETAIL_TRENDS, ADD_REGION_TRENDS,
-    ADD_CUSTOM_TREND_LOCATIONS_FILTER, CLEAR_TRENDS_SENSORS, FETCH_ANALYSIS_REGION,
-    SELECT_TRENDS_CALC_BASELINE_SETTING, SELECT_ANALYSIS_REGION, SELECT_TRENDS_REGION,
-    SELECT_TRENDS_CALC_ROLLING_SETTING, SELECT_ANALYSIS_PARAMETER, SELECT_TRENDS_PARAMETER,
-    SELECT_TRENDS_SEASON, SELECT_TRENDS_THRESHOLD, SELECT_TRENDS_VIEW_TYPE, SET_TRENDS_TIMEFRAMES,
+    ADD_ANALYSIS_COUNT, ADD_CHOOSE_TRENDS, ADD_CUSTOM_TREND_LOCATIONS_FILTER,
+    CLEAR_TRENDS_SENSORS, FETCH_ANALYSIS_REGION, SELECT_TRENDS_CALC_BASELINE_SETTING,
+    SELECT_ANALYSIS_REGION, SELECT_TRENDS_REGION, SELECT_TRENDS_CALC_ROLLING_SETTING,
+    SELECT_ANALYSIS_PARAMETER, SELECT_TRENDS_PARAMETER, SELECT_TRENDS_SEASON,
+    SELECT_TRENDS_THRESHOLD, SELECT_TRENDS_VIEW_TYPE, SET_TRENDS_TIMEFRAMES,
     SET_TRENDS_SENSORS, UPDATE_TRENDS_SENSORS
 } from '../actions';
 import {
@@ -90,22 +90,14 @@ const chosenTrends = (state:ChosenTrendsState = defaultState,
 
         case ADD_CHOOSE_TRENDS:
 
-            // push the new sensor into chosenTrends.trends_regions
-            let temp_sensor = [action.sensor];
-            let temp_data = temp_sensor.concat(state["trends_sensors"]);
+            // push the new sensor into chosenTrends.trends_sensors and
+            // update chosenTrends.trends_regions with averages
+            let tmpsensor = [action.sensor];
+            let tmpdata = tmpsensor.concat(state["trends_sensors"]);
 
             return Object.assign({}, state, {
-                trends_sensors : temp_data,
-            });
-
-        case ADD_REGION_TRENDS:
-
-            // push the new sensor into chosenTrends.trends_regions
-            let temp_region_sensor = [action.sensor];
-            let temp_region_data = temp_region_sensor.concat(state["trends_regions"]);
-
-            return Object.assign({}, state, {
-                trends_regions : temp_region_data
+                trends_sensors : tmpdata,
+                trends_regions : state.trends_regions
             });
 
         case SET_TRENDS_TIMEFRAMES: